Created task and contact routes
