+++ helperFunctions.js
@@ -18,7 +18,7 @@ hbs.registerHelper('styleSheets', (document) => {
       case 'dashboard':
       return `<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui-calendar/0.0.8/calendar.css">\n
               <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/semantic-ui@2.2.10/dist/semantic.min.css">\n
-              <link rel="stylesheet" href="stylesheets/task.css">`
+              <link rel="stylesheet" href="../stylesheets/task.css">`
       case 'task':
       return `<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui-calendar/0.0.8/calendar.css">\n
               <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/semantic-ui@2.2.10/dist/semantic.min.css">\n
