+++ routes/index.js
@@ -31,7 +31,7 @@ router.get('/', function(req, res, next) {
           lastName,
           email,
           password: hash,
-          picture: '/images/placeholder.png'
+          picture: '../images/placeholder.png'
         })
         .then((user) => {
           res.send(user.email);
