+++ views/dashboard.hbs
@@ -34,5 +34,5 @@
 </div>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/semantic-ui@2.2.10/dist/semantic.min.js"></script>
-<script src="javascripts/dashboard.js"></script>
+<script src="../javascripts/dashboard.js"></script>
