+++ views/partials/header.hbs
@@ -1,9 +1,9 @@
         <div class="tasks">
-            <a href="/dashboard/tasks" data-tasks><i class="fa fa-plus blue" aria-hidden="true"></i></a>
+            <a href="/task" data-tasks><i class="fa fa-plus blue" aria-hidden="true"></i></a>
         </div>
         <div class="contacts">
-            <a href="/dashboard/contacts" data-contacts><i class="fa fa-address-book-o blue" aria-hidden="true"></i></a>
+            <a href="/contacts" data-contacts><i class="fa fa-address-book-o blue" aria-hidden="true"></i></a>
         </div>
         <div class="settings">
-            <a href="/dashboard/settings" data-settings><i class="fa fa-cog gray" aria-hidden="true"></i></a>
+            <a href="/settings" data-settings><i class="fa fa-cog gray" aria-hidden="true"></i></a>
         </div>
